+++ src/pages/Onboarding/Onboarding.module.scss
@@ -1,0 +1,6 @@
+.signupContainer {
+  display: flex;
+  flex-direction: column;
+  width: 400px;
+  gap: 24px;
+}
