+++ src/pages/Onboarding/Steps/PersonalDetails/PersonalDetails.module.scss
