+++ src/pages/Onboarding/Steps/PersonalDetails/index.tsx
