+++ src/pages/Onboarding/Steps/StepProps.tsx
