+++ src/pages/Onboarding/index.tsx
@@ -1,0 +1,1 @@
+export { Onboarding } from './Onboarding';
