+++ src/pages/Signup/Signup.module.scss
@@ -4,3 +4,48 @@
   width: 400px;
   gap: 24px;
 }
+
+.text {
+  color: hsl(var(--primary));
+}
+
+.caption {
+  color: hsl(var(--muted-foreground));
+}
+
+.header {
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  gap: 8px;
+}
+
+.body {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.footer {
+  display: flex;
+  flex-direction: row;
+  margin: auto;
+  text-align: center;
+  gap: 4px;
+}
+
+.footerButton {
+  padding: 0;
+  height: 24px;
+}
+
+.separatorContainer {
+  display: flex;
+  flex-direction: row;
+  gap: 8px;
+}
+
+.separator {
+  width: 185px;
+  margin: auto;
+}
