--- src/pages/Signup/Steps/Register/Register.module.scss
@@ -1,44 +1,0 @@
-.text {
-  color: hsl(var(--primary));
-}
-
-.caption {
-  color: hsl(var(--muted-foreground));
-}
-
-.header {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  gap: 8px;
-}
-
-.body {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.footer {
-  display: flex;
-  flex-direction: row;
-  margin: auto;
-  text-align: center;
-  gap: 4px;
-}
-
-.footerButton {
-  padding: 0;
-  height: 24px;
-}
-
-.separatorContainer {
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-}
-
-.separator {
-  width: 185px;
-  margin: auto;
-}
