--- src/pages/Signup/Steps/Register/index.tsx
@@ -1,1 +1,0 @@
-export { Register } from './Register';
